+++ .github/release-notes.md
@@ -1,1 +1,2 @@
-The initial release of TFView.
+* Hide spectator panel by default
+* Fixed some popups breaking when opened within a popup